generator and async generator support
fixes #38
First thing to note about this is that I broke the microbundle build.
I'll attempt to fix the generator function to not use things microbundle
doesn't expect, but I'm not sure how far I'll get
with that. Which that's also something tying up this pull request:
https://github.com/developit/greenlet/pull/50

One other thing to note about this is that I decided on
returning a promise of the asyncIterator from the generator function.
This is due to the fact that the function from the user only ever gets
evaluated as an executable function on the worker side.